+++ .flake8
@@ -1,6 +1,6 @@
 [flake8]
 select = B,B9,C,D,DAR,E,F,N,RST,S,W
-ignore = E203,E501,RST201,RST203,RST301,W503
+ignore = E203,E501,RST201,RST203,RST301,W503,N815
 max-line-length = 120
 max-complexity = 10
 docstring-convention = google
